feat: album on track click, pop up window on RMC, click behavior for playlists and albums
