+++ crates/workspace/src/lib.rs
@@ -6,12 +6,14 @@ mod sidebar_left;
 mod sidebar_right;
 mod title_bar;
 mod toolbar;
+mod track_menu;
 
 pub use chrome::Chrome;
 pub use player_bar::PlayerBar;
 pub use sidebar_left::Sidebar;
 pub use title_bar::TitleBar;
 pub use toolbar::{Columned, Filterable, Searchable, Toolbar, Tooled};
+pub use track_menu::TrackMenu;
 
 use gpui::prelude::*;
 use gpui::{AnyView, App, Context, Entity, FocusHandle, Render};
